backend/requirements.txt
asgiref==3.7.2
autopep8==2.0.4
Babel==2.15.0
Django==4.2
django-cors-headers==4.3.1
django-filter==24.2
django-phonenumber-field==6.3.0
djangorestframework==3.15.1
flake8==7.0.0
gunicorn==20.1.0
load-dotenv==0.1.0
mccabe==0.7.0
phonenumbers==8.13.36
pillow==10.3.0
psycopg2==2.9.7
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
sqlparse==0.4.4
tzdata==2024.1
